requirements.txt
black==22.12.0
click==8.1.3
flake8==6.0.0
isort==5.11.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
tomli==2.0.1
